fix: expose Runtime API metadata
